Handle path separators in nested dir situations
When the "next file to play" is saved on e.g. Windows (with backslash
directory separators), it will not be found on Linux, because the file
will say "S1\E1" rather than "S1/E1".

To fix this, all filenames (paths, really) will have all their
slashes (forward and backward) replaced with the ones that work on the
current platform.