Fix running Macos apps directly
On Maocs the app from the commandline was working fine, but when you
tried to run `pls.app` from the Finder or installed via `fbs
installer` in Applications, the icon would flash briefly and then
close.

Turns out, our Windowed/CLI detection (which looked at the length of
`sys.argv`) didn't work since Macos was passing an extra command line
argument to pls.

Since we're not providing the commansline interface yet, we just
always open the window for now.

When we do, we'll need to figure this out. I think we'll just have to
create a separate file for the CLI application. Unless there's a way
to detect this somehow.