Configure the desired layout
There are going to be tweaks here and there, but this mimic's the
Python's layout exactly.

It is a bit hacky: having to use a bottom-up main layout and some of
the extra nested ones. But it's the only way I could get it to work
the way I wanted -- having all the buttons bottom-aligned and the list
of shows taking the rest of the space.