Increase the font sizes
The fonts for the "heading" and "button" styles (which covers most if
not all of the displayed text) are now bigger.

I'm still not thrilled about the rendering though. The shapes are thin
and wispy compared to the python version. That's been one of the
reasons for increasing the size tbh.

Not sure that's something we can do anything about. I'll try picking
up a different font and see if it helps.