Package the Mac OS app bundle in the deploy script
This is a pretty big change to the deploy script. It now creates both
zipfiles and tarballs using the relevant Rust crates (tar and zip),
it's able to package the .app bundle (which is secretly a directory)
and we take all the arguments form the environment rather than doing
platform detection logic in the script.

The code could use some cleanup, but it does in fact seem to do
everything that's currently needed.